Whitelisting: allow non-whitelisting rules
Allow 'whitelist' to include also non-whitelisting rules. This makes it
possible to use the 'whitelist' as a more generic record of manual
analysis results even if the analysed vulnerabilities should not be
whitelisted.
