+++ sbomnix/utils.py
@@ -105,76 +105,6 @@ def df_log(df, loglevel, tablefmt="presto"):
         LOG.log(loglevel, "\n%s\n", table)
 
 
-def load_vuln_whitelist(whitelist_csv):
-    """
-    Load vulnerability whitelist from whitelist_csv. Returns None
-    if whitelist_csv is not a valid vulnerability whitelist. Otherwise
-    returns whitelist_csv as dataframe
-    """
-    try:
-        df = df_from_csv_file(whitelist_csv)
-        # Whitelist must have the following columns
-        if not set(["vuln_id", "comment"]).issubset(df.columns):
-            LOG.warning("Whitelist csv missing required columns")
-            return None
-        return df
-    except pd.errors.ParserError:
-        return None
-
-
-def df_apply_vuln_whitelist(df_whitelist, df_vulns):
-    """
-    Apply df_whitelist to vulnerabilities in df_vulns, changing df_vulns
-    in-place.
-    Adds columns "whitelist" and "whitelist_comment" to df_vulns based
-    on whitelisting regular expressions in df_whitelist["vuln_id"].
-    If df_whitelist["package"] exists and is not empty, require strict
-    match in df_whitelist["package"] and df_vulns["package"].
-    """
-    # Add default values to whitelist columns
-    df_vulns["whitelist"] = False
-    df_vulns["whitelist_comment"] = ""
-    if "vuln_id" not in df_vulns:
-        LOG.fatal("Missing 'vuln_id' column from df_vulns")
-        sys.exit(1)
-    if "vuln_id" not in df_whitelist:
-        LOG.warning("Whitelist ignored: missing 'vuln_id' column from whitelist")
-        return
-    check_pkg_name = False
-    if "package" in df_whitelist.columns and "package" in df_vulns.columns:
-        check_pkg_name = True
-    # Iterate rows in df_whitelist in reverse order so the whitelist rules
-    # on top of the file get higher priority
-    df_whitelist_rev = df_whitelist[::-1]
-    for whitelist_entry in df_whitelist_rev.itertuples():
-        LOG.log(LOG_SPAM, "whitelist_entry: %s", whitelist_entry)
-        regex = str(whitelist_entry.vuln_id).strip()
-        LOG.log(LOG_SPAM, "whitelist regex: %s", regex)
-        df_matches = df_vulns["vuln_id"].str.fullmatch(regex)
-        if check_pkg_name and whitelist_entry.package:
-            LOG.log(LOG_SPAM, "filtering by pacakge name: %s", whitelist_entry.package)
-            df_matches = df_matches & (df_vulns["package"] == whitelist_entry.package)
-        df_vulns.loc[df_matches, "whitelist"] = True
-        df_vulns.loc[df_matches, "whitelist_comment"] = whitelist_entry.comment
-        LOG.log(LOG_SPAM, "matches %s vulns", len(df_vulns[df_matches]))
-        df_log(df_vulns[df_matches], LOG_SPAM)
-
-
-def df_drop_whitelisted(df):
-    """
-    Drop whitelisted vulnerabilities from `df` as well as
-    the related columns.
-    """
-    if "whitelist" in df.columns:
-        # Convert possible string to boolean
-        df["whitelist"] = df["whitelist"].replace({"True": True, "False": False})
-        df = df[~df["whitelist"]]
-        df = df.drop("whitelist", axis=1)
-    if "whitelist_comment" in df.columns:
-        df = df.drop("whitelist_comment", axis=1)
-    return df
-
-
 def exec_cmd(cmd, raise_on_error=True, return_error=False):
     """Run shell command cmd"""
     command_str = " ".join(cmd)
